feat: support cucumber regex step syntax
